:broom: relocates transactions from inititalizer file
Issue:
- https://github.com/samvera/bulkrax/issues/897
