+++ lib/bulkrax/engine.rb
@@ -17,6 +17,7 @@ module Bulkrax
       require 'bulkrax/persistence_layer'
       require 'bulkrax/persistence_layer/active_fedora_adapter' if defined?(ActiveFedora)
       require 'bulkrax/persistence_layer/valkyrie_adapter' if defined?(Valkyrie)
+      require 'bulkrax/transactions' if defined?(Valkyrie)
     end
 
     config.generators do |g|
