+++ lib/generators/bulkrax/templates/config/initializers/bulkrax.rb
@@ -86,98 +86,3 @@ end
 if Object.const_defined?(:Hyrax) && ::Hyrax::DashboardController&.respond_to?(:sidebar_partials)
   Hyrax::DashboardController.sidebar_partials[:repository_content] << "hyrax/dashboard/sidebar/bulkrax_sidebar_additions"
 end
-
-# TODO: move outside of initializer?
-class BulkraxTransactionContainer
-  extend Dry::Container::Mixin
-
-  namespace "work_resource" do |ops|
-    ops.register "create_with_bulk_behavior" do
-      steps = Hyrax::Transactions::WorkCreate::DEFAULT_STEPS.dup
-      steps[steps.index("work_resource.add_file_sets")] = "work_resource.add_bulkrax_files"
-
-      Hyrax::Transactions::WorkCreate.new(steps: steps)
-    end
-
-    ops.register "update_with_bulk_behavior" do
-      steps = Hyrax::Transactions::WorkUpdate::DEFAULT_STEPS.dup
-      steps[steps.index("work_resource.add_file_sets")] = "work_resource.add_bulkrax_files"
-
-      Hyrax::Transactions::WorkUpdate.new(steps: steps)
-    end
-
-    # TODO: uninitialized constant BulkraxTransactionContainer::InlineUploadHandler
-    # ops.register "add_file_sets" do
-    #   Hyrax::Transactions::Steps::AddFileSets.new(handler: InlineUploadHandler)
-    # end
-
-    ops.register "add_bulkrax_files" do
-      Bulkrax::Transactions::Steps::AddFiles.new
-    end
-  end
-end
-Hyrax::Transactions::Container.merge(BulkraxTransactionContainer)
-
-# TODO: move outside of initializer?
-module HasMappingExt
-  ##
-  # Field of the model that can be supported
-  def field_supported?(field)
-    field = field.gsub("_attributes", "")
-
-    return false if excluded?(field)
-    return true if supported_bulkrax_fields.include?(field)
-
-    property_defined = factory_class.singleton_methods.include?(:properties) && factory_class.properties[field].present?
-
-    factory_class.method_defined?(field) && (Bulkrax::ValkyrieObjectFactory.schema_properties(factory_class).include?(field) || property_defined)
-  end
-
-  ##
-  # Determine a multiple properties field
-  def multiple?(field)
-    @multiple_bulkrax_fields ||=
-      %W[
-        file
-        remote_files
-        rights_statement
-        #{related_parents_parsed_mapping}
-        #{related_children_parsed_mapping}
-      ]
-
-    return true if @multiple_bulkrax_fields.include?(field)
-    return false if field == "model"
-
-    field_supported?(field) && (multiple_field?(field) || factory_class.singleton_methods.include?(:properties) && factory_class&.properties&.[](field)&.[]("multiple"))
-  end
-
-  def multiple_field?(field)
-    form_definition = schema_form_definitions[field.to_sym]
-    form_definition.nil? ? false : form_definition.multiple?
-  end
-
-  # override: we want to directly infer from a property being multiple that we should split when it's a String
-  # def multiple_metadata(content)
-  #   return unless content
-
-  #   case content
-  #   when Nokogiri::XML::NodeSet
-  #     content&.content
-  #   when Array
-  #     content
-  #   when Hash
-  #     Array.wrap(content)
-  #   when String
-  #     String(content).strip.split(Bulkrax.multi_value_element_split_on)
-  #   else
-  #     Array.wrap(content)
-  #   end
-  # end
-
-  def schema_form_definitions
-    @schema_form_definitions ||= ::SchemaLoader.new.form_definitions_for(factory_class.name.underscore.to_sym)
-  end
-end
-[Bulkrax::HasMatchers, Bulkrax::HasMatchers.singleton_class].each do |mod|
-  mod.prepend HasMappingExt
-end
